requirements.txt
matplotlib>=3.3.1
numpy>=1.17.3
Pillow>=6.1.0

Cython>=0.19.2
numpy>=1.7.1
scipy>=0.13.2
opencv-python

django
djangorestframework
django-cors-headers